Renamed SyntheticBackPressureTest to SyntheticTest since we use it for a lot of purposes.
SyntheticOperations now implement getservicename to be correct with the servicename removal in hz serialization:
https://github.com/hazelcast/hazelcast/pull/4468
